+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/preprocessing/StandardScaler.java
@@ -129,9 +129,9 @@ public class StandardScaler extends AbstractScaler<StandardScaler.ModelParameter
         Map<Object, Double> stdColumnValues = modelParameters.getStdColumnValues();
         boolean scaleResponse = knowledgeBase.getTrainingParameters().getScaleResponse();
 
-        Set<TypeInference.DataType> supportedTypes = getSupportedTypes();
+        Set<TypeInference.DataType> supportedXDataTypes = getSupportedXDataTypes();
         Stream<Object> transformedColumns = trainingData.getXDataTypes().entrySet().stream()
-                .filter(e -> supportedTypes.contains(e.getValue()))
+                .filter(e -> supportedXDataTypes.contains(e.getValue()))
                 .map(e -> e.getKey());
 
         streamExecutor.forEach(StreamMethods.stream(transformedColumns, isParallelized()), column -> {